+++ src/data/cardItemsList.js
